refactor: window 패키지 구현
